✨ feat(home): consolidate search book functionality in 'Home' tab
- remove 'search-book' tab
- integrate search book functionality into 'Home' tab
- create new 'BookSearchResult' component for search results
- delete 'search-book.tsx' component file
